# [FIX] stock: speed up _find_delivery_ids_by_lot