Rewrite `_find_delivery_ids_by_lot` to reduce the number
of recursive calls by grouping the producing move lines
together before each recursive call. This reduces
the number of search calls, speeding up the whole method.

The lines 226-228 are not directly related to the optimization of the method.
PR odoo/odoo#81975 introduced an unwanted "bug" by solving a possible traceback.
Currently, the position of a given lot in `self` impacts the result. In the customer DB
for instance, the lot 4283 had 67 related picking_ids if it was at the beginning of self
and 1393 picking_ids when at the end. This is because simply passing `delivery_by_lot`
as a param sets each lot picking_ids to the ids of all the previously processed lots. 
Adding a condition on lot_path prevents this issue.
 
A unittest has been added that reproduces the scenario of PR odoo/odoo#81975 to ensure
that the new fix still prevents the occurrence of a traceback.

#### Speed up

The results presented here are a bit messy because the computing time can change drastically
between two lots, even more so when working on recordsets. The customer had 7000 lots 
and 32000 pickings. The first result set was obtained by averaging the computation time
over different recordsets. The second set was obtained by averaging the computation with a
fixed recordset (so the 10 lots recordset was a subset of the 80 lots one).

| # Lots | Before PR | After PR |
|:------:|:----------:|:---------:|
| 1 | 0.44s | 0.06s |
| 10 | 3.64s | 0.42s |
| 25 | 94s | 12s |
| 50 | 86s | 10.98s |
| 80 | 85s | 10.5s |

| # Lots | Before PR | After PR |
|:------:|:-----------:|:--------:|
| 1 | 1.05s | 0.1s |
| 10 | 6.5s | 0.61s |
|  25 | 9.5s | 1.27s |
|  50 | 21s | 2.32s |
|  80 | 37.5s | 3.85s |

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr