@robodoo 2022-06-22
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/94320).