@amoyaux 2022-06-30
https://github.com/odoo/odoo/pull/94775